+++ packages/tspd/src/ref-doc/components/typekits-file.tsx
@@ -43,7 +43,9 @@ async function flattenOutput(output: OutputDirectory): Promise<Record<string, st
   });
 
   for (const file of rawFiles) {
-    files[file.path] = await format(file.contents);
+    if ("contents" in file) {
+      files[file.path] = await format(file.contents);
+    }
   }
   return files;
 }