+++ packages/http-client-python/CHANGELOG.md
@@ -1,5 +1,61 @@
 # Change Log - @typespec/http-client-python
 
+## 0.15.2
+
+### Bug Fixes
+
+- [#8175](https://github.com/microsoft/typespec/pull/8175) fix generated output folder for `generated_samples/generated_tests`
+- [#8117](https://github.com/microsoft/typespec/pull/8117) keep declaration of pyproject.toml same with existing setup.py of ARM SDK
+
+
+## 0.15.1
+
+### Bump dependencies
+
+- [#7968](https://github.com/microsoft/typespec/pull/7968) Adopt latest TCGC.
+- [#8135](https://github.com/microsoft/typespec/pull/8135) Bump tsp packages to 1.3.0 and 0.59.0
+
+### Bug Fixes
+
+- [#7911](https://github.com/microsoft/typespec/pull/7911) [python] don't fail on response body consumption in `_failsafe_deserialize`
+- [#8124](https://github.com/microsoft/typespec/pull/8124) Don't include folder suffixes in documentation generated with `generation-subdir`
+- [#8114](https://github.com/microsoft/typespec/pull/8114) Exclude doc folder in pyproject.toml
+
+
+## 0.15.0
+
+### Features
+
+- [#7829](https://github.com/microsoft/typespec/pull/7829) Adding pyproject.toml generation and optional keep-setup-py flag
+- [#7994](https://github.com/microsoft/typespec/pull/7994) add `generation-subdir` flag
+
+### Bug Fixes
+
+- [#8070](https://github.com/microsoft/typespec/pull/8070) fix import for error model in multi namespace
+- [#8091](https://github.com/microsoft/typespec/pull/8091) Don't hardcode `emit-cross-language-definition-file` as `true` for azure generations
+- [#8006](https://github.com/microsoft/typespec/pull/8006) Ensure necessary typing imports for internal models
+
+
+## 0.14.2
+
+### Bug Fixes
+
+- [#8058](https://github.com/microsoft/typespec/pull/8058) fix outputfolder of packaging files for arm sdk
+
+
+## 0.14.1
+
+### Bump dependencies
+
+- [#7820](https://github.com/microsoft/typespec/pull/7820) Bump min dep of generated sdks on `azure-core` to `1.35.0` for backcompat serialization methods
+
+### Bug Fixes
+
+- [#7992](https://github.com/microsoft/typespec/pull/7992) Import mixins from operations init file to get patch changes
+- [#8039](https://github.com/microsoft/typespec/pull/8039) Don't hardcode client in sample to first client in list
+- [#7939](https://github.com/microsoft/typespec/pull/7939) Pad `datetime` as a model property for typespec
+
+
 ## 0.14.0
 
 ### Features